correcciones esperas explicitas
