+++ src/page/bank-payment.page.ts
@@ -4,14 +4,15 @@ export class BankPaymentPage {
   private proceedToCheckoutButton: ElementFinder;
 
   constructor () {
-    this.proceedToCheckoutButton = $('[style*="display: block;"] .button-container > a');
+    // tslint:disable-next-line: max-line-length
+    this.proceedToCheckoutButton = $('a.btn.btn-default.button.button-medium span');
   }
 
   private EC = protractor.ExpectedConditions;
 
   public async proceedToCheckout(): Promise<void> {
     const isClickable = this.EC.elementToBeClickable(this.proceedToCheckoutButton);
-    browser.wait(isClickable, 5000);
+    await browser.wait(isClickable, 2000);
     await this.proceedToCheckoutButton.click();
   }
 }
